[REFACTOR] Update token description validation related code
Description filters are separated from the tokengroup manager.
As they are put into a vector of std::function's, they can be made switched on and off at run-time.

Before the description object is databased, it is parsed and truncated to the maximum length.
Next, the filters are applied.
And finally, they are added to the token manager - and thus made available.

Filtering actions are captured in a status object.